Add additional casts for int64_t printf()s
The recent addition of printf() string checking to vboot has shown up errors
for sandbox.

Unfortunately sandbox does not seem coherence on this point. The uint64_t is
likely just 'unsigned long' but the printf() formation string for PRIu64 is
something else, so we get a warning.

Adjust the code to make sure it builds on ARM and sandbox.

BUG=none
BRANCH=none
TEST=manual
crosfw -b sandbox -V
crosfw -b peach_pit -V

See there are no warnings in either case.

+++ firmware/lib/vboot_audio.c
@@ -222,7 +222,8 @@ VbAudioContext *VbAudioOpen(LoadKernelParams *lkparams)
 	VbExSleepMs(10);
 	b = VbExGetTimer();
 	ticks_per_msec = (b - a) / 10ULL ;
-	VBDEBUG(("VbAudioOpen() - ticks_per_msec is %llu\n", ticks_per_msec));
+	VBDEBUG(("VbAudioOpen() - ticks_per_msec is %llu\n",
+		 (unsigned long long)ticks_per_msec));
 
 	/* Initialize */
 	Memset(audio, 0, sizeof(*audio));
